Add header section
Heading is created based on TOP's guide. One useful thing I learned
is to utilize 'min-width' to restrict movements of elements and
prevent containers' overflow when browser is shrunk. In this commit,
this technique is used to prevent 'header-content-left' from
overflowing, and to restrict the links (top right) from sliding around.